requirements.txt
tika
python-pdfbox